Cleaned ping service (separate files)
