+++ src/modules/interface/tunnel_config.h
@@ -1,0 +1,17 @@
+#ifndef __TUNNELCONFIG_H
+#define __TUNNELCONFIG_H
+
+#include <stdbool.h>
+#include <stdint.h>
+#include <string.h>
+
+// Returns the number of drones to be considered in the chain
+uint8_t getNDrones();
+
+// Sets the number of drones in the chain
+void setNDrones(uint8_t ndrones);
+
+// Get the last 4 bits of this drone's address
+uint8_t getDroneId();
+
+#endif
