+++ src/modules/src/tunnel_config.c
@@ -1,0 +1,18 @@
+#include "tunnel_config.h"
+
+#include "configblock.h"
+
+// Number of available drones in the tunnel exploration crew
+static uint8_t nDrones = 3; // 15 max because of the 4-bit addresses
+
+uint8_t getNDrones() {
+    return nDrones;
+}
+
+void setNDrones(uint8_t ndrones) {
+    nDrones = ndrones;
+}
+
+uint8_t getDroneId() {
+  return (uint8_t)configblockGetRadioAddress() & 0x0F;
+}
