+++ src/utils/interface/configblock.h
@@ -37,7 +37,6 @@ bool configblockTest(void);
 int configblockGetRadioChannel(void);
 int configblockGetRadioSpeed(void);
 uint64_t configblockGetRadioAddress(void);
-uint8_t configblockGetRadioAddressShort(void);
 
 float configblockGetCalibPitch(void);
 float configblockGetCalibRoll(void);
